Implement `Error` for `NotmuchStatus`.
Entails implementation of `ToStr` and `Display`, also.

Note that the `to_str()` method uses the `to_static_str()` method on the
`ToStaticStr` trait internally --- this is how I expect to use the
latter.  This lets me provide a uniform `to_str()` method in the rest of
the API, while transparently handling the differences between C string
lifetimes.