Make `to_str()` method generic over lifetime.
If I have this correct:

The output `&str` will have the same lifetime as whatever the lifetime
of the block in which it is called.

This allows using it one the output of notmuch FFI calls, by propogating
the lifetime of the safe wrapper object (e.g., `Database`) to the `&str`
returned from `to_str()`.  This mirrors the lifetime of the actual
underlying C string.

If I don't grok lifetimes as well as I think, this could be a lie.

Future self: you should add unit tests to test the above explanation.