Make methods borrow self mutably when appropriate.
Use of raw pointers and FFI functions mean that the compiler won't force
us to be correct in our usage of `&self` vs. `&mut self`.

These functions make changes to the notmuch database, and we want the
type system to ensure we can't simultaneously manipulate the database
from different functions.