Fix test environment leakage
The generated config files for the test maildirs did not contain the
`hook_dir` option, and notmuch surprisingly defaulted to using
my *actual* hooks when running the test suite. This patch fixes this by
adding the `hook_dir` option to the generated config file, and also
removes a number of other environment variables that the `notmuch`
command could potentially pick up on certain systems.